Added generated route on TrekRouteScreen.

+++ src/backend/server.py
@@ -7,7 +7,7 @@ app = Flask(__name__)
 # API given coordinates and distance to generate randomized route
 @app.route('/route/<lat>/<lon>/<d>')
 def route(lat, lon, d):
-    new_lat = str(int(lat) + int(d) / 69)
+    new_lat = str(float(lat) + float(d) / 69)
     url = "https://maps.googleapis.com/maps/api/directions/json?origin=" + lat + "," + lon + "&destination=" + lat + "," + lon + "&mode=walking"+ "&waypoints=" + new_lat + "," + lon + "&key=AIzaSyAai4YqEuIUOKHQl8i1LgPcM6JEjyIFuuM"
     response_dict = requests.get(url).json()
     return response_dict
